Further work on the LMC emulator

+++ src/chips/trivial/lmc/instructions/add.cpp
@@ -7,16 +7,16 @@ namespace emu::lmc {
 
 
     /**
-     * Add from memory to accumulator
+     * Add value in memory to the accumulator
      * <ul>
-     *   <li>Size: 1</li>
+     *   <li>Size: 2</li>
      * </ul>
      *
      * @param acc_reg is the accumulator register, which will be mutated
-     * @param address is the value to add to the accumulator register
+     * @param address is the address to the value in memory
      * @param memory is the memory
      */
-    void add(u8 &acc_reg, u8 address, const EmulatorMemory<u16, u8> &memory) {
+    void add(u16 &acc_reg, u8 address, const EmulatorMemory<u8, u16> &memory) {
         acc_reg += memory.read(address);
     }
 }
